+++ src/components/msg/group/create_group_request/index.tsx
@@ -1,0 +1,35 @@
+import React from 'react';
+import Trans from 'next-translate/Trans';
+import { Typography } from '@material-ui/core';
+import { Name } from '@components';
+import { MsgCreateGroupRequest } from '@models';
+import {
+  useProfileRecoil,
+} from '@recoil/profiles';
+
+const CreateGroupRequest = (props: {
+  message: MsgCreateGroupRequest;
+}) => {
+  const { message } = props;
+
+  const admin = useProfileRecoil(message.admin);
+  const adminMoniker = admin ? admin?.name : message.admin;
+
+  return (
+    <Typography>
+      <Trans
+        i18nKey="message_contents:MsgCreateGroupRequest"
+        components={[
+          (
+            <Name
+              address={message.admin}
+              name={adminMoniker}
+            />
+          ),
+        ]}
+      />
+    </Typography>
+  );
+};
+
+export default CreateGroupRequest;