+++ src/components/msg/group/update_group_account_metadata_request/index.test.tsx
@@ -1,0 +1,50 @@
+import React from 'react';
+import { RecoilRoot } from 'recoil';
+import renderer from 'react-test-renderer';
+import { MockTheme } from '@tests/utils';
+import { MsgUpdateGroupAccountMetadataRequest } from '@models';
+import UpdateGroupAccountMetadataRequest from '.';
+
+// ==================================
+// mocks
+// ==================================
+
+jest.mock('@components', () => ({
+  Name: (props) => <div id="Name" {...props} />,
+}));
+
+jest.mock('next-translate/Trans', () => (
+  (props) => <div id="Trans" {...props} />
+));
+
+// ==================================
+// unit tests
+// ==================================
+describe('screen: TransactionDetails/UpdateGroupAccountMetadataRequest', () => {
+  it('matches snapshot', () => {
+    const message = new MsgUpdateGroupAccountMetadataRequest({
+      category: 'group',
+      type: 'MsgUpdateGroupAccountMetadataRequest',
+      admin: 'admin',
+      address: 'address',
+    });
+
+    const component = renderer.create(
+      <RecoilRoot>
+        <MockTheme>
+          <UpdateGroupAccountMetadataRequest
+            message={message}
+          />
+        </MockTheme>
+      </RecoilRoot>,
+    );
+    const tree = component.toJSON();
+    expect(tree).toMatchSnapshot();
+
+    expect(component.root.findByProps({ id: 'Trans' }).props.i18nKey).toEqual('message_contents:MsgUpdateGroupAccountMetadataRequest');
+  });
+
+  afterEach(() => {
+    jest.clearAllMocks();
+  });
+});