+++ src/components/msg/group/vote_request/index.tsx
@@ -1,0 +1,35 @@
+import React from 'react';
+import Trans from 'next-translate/Trans';
+import { Typography } from '@material-ui/core';
+import { Name } from '@components';
+import { MsgVoteRequest } from '@models';
+import {
+  useProfileRecoil,
+} from '@recoil/profiles';
+
+const VoteRequest = (props: {
+  message: MsgVoteRequest;
+}) => {
+  const { message } = props;
+
+  const voter = useProfileRecoil(message.voter);
+  const voterMoniker = voter ? voter?.name : message.voter;
+
+  return (
+    <Typography>
+      <Trans
+        i18nKey="message_contents:MsgVoteRequest"
+        components={[
+          (
+            <Name
+              address={message.voter}
+              name={voterMoniker}
+            />
+          ),
+        ]}
+      />
+    </Typography>
+  );
+};
+
+export default VoteRequest;